+++ src/lib/auth.ts
@@ -14,7 +14,6 @@ export const authOptions: NextAuthOptions = {
     callbacks: {
         async session({ session, token }) {
             if (session.user) {
-                // @ts-ignore
                 session.user.id = token.sub;
             }
             return session;